Moved state.py and component.py one level up, and associated changes
