+++ src/mepo/command/branch_create.py
@@ -1,4 +1,4 @@
-from ..state.state import MepoState
+from ..state import MepoState
 from ..utilities import verify
 from ..git import GitRepository
