+++ src/mepo/command/reset.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 
-from ..state.state import MepoState
+from ..state import MepoState
 from ..utilities.exceptions import NotInRootDirError
 
 from .clone import run as mepo_clone_run
