+++ src/mepo/command/stash_list.py
@@ -1,4 +1,4 @@
-from ..state.state import MepoState
+from ..state import MepoState
 from ..git import GitRepository
 
 def run(args):
