+++ src/mepo/component.py
@@ -5,9 +5,9 @@ import textwrap
 from collections import namedtuple
 from urllib.parse import urlparse
 
-from ..utilities import shellcmd
-from ..utilities import mepoconfig
-from ..utilities.version import MepoVersion
+from .utilities import shellcmd
+from .utilities import mepoconfig
+from .utilities.version import MepoVersion
 
 # This will be used to store the "final nodes" from each subrepo
 original_final_node_list = []
