+++ src/mepo/state.py
@@ -7,10 +7,10 @@ import pickle
 from pathlib import Path
 
 from .component import MepoComponent
-from ..utilities.exceptions import StateDoesNotExistError
-from ..utilities.exceptions import StateAlreadyInitializedError
+from .utilities.exceptions import StateDoesNotExistError
+from .utilities.exceptions import StateAlreadyInitializedError
 
-from ..utilities import shellcmd
+from .utilities import shellcmd
 
 class MepoState(object):
